Add Swag Reminder Email + Email Refresh (#95)
* Add Swag Reminder Email

Adds automated email for MAGWest swag reminder. Based on super's, except we don't have a deadline, so changed the when condition to basically be 2 days before the start of the event.

* Add attendee_swag_promo.html for email.

Adding a custom email with proper variables using Super's as a template.

* Refresh volunteer_food.txt

Reflects 2019 specific language/processes.

* Update guest_food_info.txt

Reflects 2019 specific info/processes.

* Update prefest-faq

Updates 2019 relevant information.

* Update prefest_faq.html

Additional updates to FAQ copy.

* Final edits to email.

* Last minute change to FAQ

* Fix scheduled email syntax
Now it works!

* Update prefest FAQ to use URL_BASE
URL_BASE is what we use in all our other emails, so we want to use it here to make sure the generated link is correct.